// fe-angular/src/app/configs/models/route.enum.ts
export enum RoutePath {
  LANDING = 'home',
  DASHBOARD = 'dashboard',
  AGENT_MANAGEMENT = 'agent-management',
  AGENTS_DETAIL = 'agent-detail',
  AGENTS_NEW = 'agent-create',
  LOGS = 'logs',
  LOGS_DETAIL = 'detail',
  CONFIGURATION = 'configuration',
  USER_MANEGEMENT = 'user-management',
  USER_LOGIN = 'login',
  USER_REGISTER = 'register',
  USER_FORGOT_PASSWORD = 'forgot-password',
  EMPTY = '',
  JOBS = 'jobs',
  JOB_MANAGEMENT = 'job-management',
  JOBS_MONITOR = 'job-monitor',
  POST = 'post',
  DATA_MANAGEMENT = 'data-management',
  PROCESSING_LOG = 'video-logs',
}

export enum ROUTE_TYPE {
  TOP = 'TOP',
  CONFIGURATION = 'CONFIGURATION',
}